pyproject.toml
[project]
name = "agentic-development"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-genai>=1.25.0",
    "jinja2>=3.1.6",
    "langgraph>=0.5.3",
    "pandas>=2.3.1",
    "pyodb>=0.1.5",
    "pyodbc>=5.2.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "tabulate>=0.9.0",
    "twilio>=9.6.5",
    "uvicorn>=0.35.0",
]
